% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{financials}
\alias{financials}
\title{financials}
\format{
Tibble with 505 features and 14 fields
}
\usage{
financials
}
\description{
Dataset from Kaggle: https://www.kaggle.com/paytonfisher/sp-500-companies-with-financial-information
}
\keyword{datasets}
